[ADD] Options authorization and exception handling

+++ src/options/options.service.ts
@@ -1,4 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
+import { User, UserRole } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateOptionDto } from './dto/create-option.dto';
 import { UpdateOptionDto } from './dto/update-option.dto';
@@ -7,32 +12,64 @@ import { UpdateOptionDto } from './dto/update-option.dto';
 export class OptionsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  create(createOptionDto: CreateOptionDto) {
-    return this.prisma.option.create({
-      data: createOptionDto,
-    });
+  create(user: User, createOptionDto: CreateOptionDto) {
+    if (user.role !== UserRole.TEACHER) {
+      throw new ForbiddenException("You are not allowed to access this resource!");
+    }
+
+    try {
+      return this.prisma.option.create({
+        data: createOptionDto,
+      }); 
+    } catch (err) {
+      throw new BadRequestException("Can't create option!", err.message);
+    }
   }
 
   findAll() {
-    return this.prisma.option.findMany();
+    try {
+      return this.prisma.option.findMany();  
+    } catch (err) {
+      throw new BadRequestException("Can't fetch option!", err.message);
+    }
   }
 
   findOne(id: number) {
-    return this.prisma.option.findUnique({
-      where: { id },
-    });
+    try {
+      return this.prisma.option.findUniqueOrThrow({
+        where: { id },
+      }); 
+    } catch (err) {
+      throw new BadRequestException("Can't fetch option!", err.message);
+    }
   }
 
-  update(id: number, updateOptionDto: UpdateOptionDto) {
-    return this.prisma.option.update({
-      where: { id },
-      data: updateOptionDto,
-    });
+  update(user: User, id: number, updateOptionDto: UpdateOptionDto) {
+    if (user.role !== UserRole.TEACHER) {
+      throw new ForbiddenException("You are not allowed to access this resource!");
+    }
+
+    try {
+      return this.prisma.option.update({
+        where: { id },
+        data: updateOptionDto,
+      }); 
+    } catch (err) {
+      throw new BadRequestException("Failed to update option!", err.message);
+    }
   }
 
-  remove(id: number) {
-    return this.prisma.option.delete({
-      where: { id },
-    });
+  remove(user: User, id: number) {
+    if (user.role !== UserRole.TEACHER) {
+      throw new ForbiddenException("You are not allowed to access this resource!");
+    }
+
+    try {
+      return this.prisma.option.delete({
+        where: { id },
+      }); 
+    } catch (err) {
+      throw new BadRequestException("Failed to delete option!", err.message);
+    }
   }
 }
